0.2.0: make it work on the beta/stable channels.
This is a [breaking-change]. Traits being mopafied formerly needed to
extend `std::any::Any`; now they need to extend `mopa::Any`.

Users of `#![no_std]` will now need to enable the `no_std` feature on
this crate.

Users of this library formerly also needed `#![feature(core)]` on the
crate; this is necessary no longer, because mopa works on Rust stable
and beta.