Replace `TraitObject` with raw pointer casts.
The two should be completely equivalent. And if that point about the
layout of fat pointers were to change (which seems unlikely) we’d need
to change code anyway. Raw pointer casts lets us get rid of the
`transmute` calls and is shorter (A diff of -14 lines? I’ll take it.)

And @bluss asked for it.

Fixes #6.